Further modularised code by changing file structure and giving distribution its own module
